Moved repl into repl file
- moved repl logic from main.rs to repl.rs
- moved error/message into associated library files
- updated make file and added run command
