+++ Makefile
@@ -3,16 +3,16 @@ main: src/main.rs librepl.rlib libprocessor.rlib libtable.rlib dirs
 	@mv main build
 	@ln -s build/main main
 
-librepl.rlib: src/repl.rs dirs
-	rustc --crate-type lib src/repl.rs
+librepl.rlib: src/repl.rs libprocessor.rlib libtable.rlib dirs
+	rustc -L build --crate-type lib src/repl.rs
 	@mv librepl.rlib build
 
-libprocessor.rlib: src/processor.rs librepl.rlib libtable.rlib dirs
+libprocessor.rlib: src/processor.rs libtable.rlib dirs
 	rustc -L build --crate-type lib src/processor.rs
 	@mv libprocessor.rlib build
 
-libtable.rlib: src/table.rs librepl.rlib dirs
-	rustc -L build --crate-type lib src/table.rs
+libtable.rlib: src/table.rs  dirs
+	rustc --crate-type lib src/table.rs
 	@mv libtable.rlib build
 
 dirs:
@@ -22,3 +22,6 @@ clean:
 	rm main || true
 	rm -r build || true
 	find -name "*~" -type f -delete
+
+run: main
+	@./main
